Rename ParticleLayoutIO to LatticeIO.
