+++ src/QMCWaveFunctions/tests/test_kspace_jastrow.cpp
@@ -19,7 +19,7 @@
 #include "QMCWaveFunctions/WaveFunctionComponent.h"
 #include "QMCWaveFunctions/Jastrow/kSpaceJastrow.h"
 #include "QMCWaveFunctions/Jastrow/kSpaceJastrowBuilder.h"
-#include "ParticleIO/ParticleLayoutIO.h"
+#include "ParticleIO/LatticeIO.h"
 
 #include <stdio.h>
 #include <string>
